use build tags to separate tests
