+++ .github/workflows/composite-actions/kubernetes-e2e-tests/action.yaml
@@ -3,37 +3,29 @@ name: Kubernetes End-to-End tests
 # These tests are located in the /test/kubernetes/e2e directory
 description: Kubernetes End-to-End tests
 
+inputs:
+  cluster-name:
+    required: true
+    description: The name of the KinD cluster
+  test-args:
+    required: true
+    description: The list of arguments passed to the test invocation
+
 runs:
   using: "composite"
   steps:
   - name: Prep Go Runner
     uses: ./.github/workflows/composite-actions/prep-go-runner
-  - name: Install kind
-    uses: helm/kind-action@v1.5.0
-    with:
-      install_only: true
-      version: ${{ matrix.tool-versions.kind }}
-  - uses: azure/setup-kubectl@v3
-    id: kubectl
-    with:
-      version: ${{ matrix.tool-versions.kubectl }}
-  - uses: azure/setup-helm@v3
-    with:
-      version: ${{ matrix.tool-versions.helm }}
-  - name: Setup test env
-    shell: bash
-    env:
-      CLUSTER_NAME: 'kind'
-      CLUSTER_NODE_VERSION: ${{ matrix.tool-versions.cluster-node }}
-    run: ./ci/kind/setup-kind.sh
   - name: Install test tools
     shell: bash
     run: make install-test-tools
   - name: Execute tests
     env:
-      GO_TEST_USER_ARGS: ${{ matrix.test-suites.go-test-args }}
-      CLUSTER_NAME: 'kind'
-      TEST_PKG: test/kubernetes/e2e/${{ matrix.test-suites.package }}
+      GO_TEST_USER_ARGS: ${{ inputs.test-args }}
+      CLUSTER_NAME: ${{ inputs.cluster-name }}
+      ## We want to move all tests into the /tests directory
+      ## In the meantime, we rely on `...` to recursively find tests to run
+      TEST_PKG: test/kubernetes/e2e/tests,test/kubernetes/e2e/k8sgateway
     shell: bash
     run: make go-test
   # TODO: Emit artifacts to shared location on failure
