+++ .github/workflows/composite-actions/setup-kind-cluster/action.yaml
@@ -1,0 +1,42 @@
+name: Setup KinD Cluster
+
+description: The action that will seed a KinD cluster.
+
+inputs:
+  cluster-name:
+    required: true
+    description: The name of the KinD cluster
+  kind-version:
+    required: true
+    description: The version of KinD
+  kind-node-version:
+    required: true
+    description: The Node version used by KinD
+  kubectl-version:
+    required: true
+    description: The version of Kubectl
+  helm-version:
+    required: true
+    description: The version of Helm
+
+runs:
+  using: "composite"
+  steps:
+    - name: Install kind
+      uses: helm/kind-action@v1.5.0
+      with:
+        install_only: true
+        version: ${{ inputs.kind-version }}
+    - uses: azure/setup-kubectl@v3
+      id: kubectl
+      with:
+        version: ${{ inputs.kubectl-version }}
+    - uses: azure/setup-helm@v3
+      with:
+        version: ${{ inputs.helm-version }}
+    - name: Setup test env
+      shell: bash
+      env:
+        CLUSTER_NAME: ${{ inputs.cluster-name }}
+        CLUSTER_NODE_VERSION: ${{ inputs.kind-node-version }}
+      run: ./ci/kind/setup-kind.sh
