+++ test/kubernetes/e2e/k8sgateway/automtls_istio_test.go
@@ -1,3 +1,5 @@
+//go:build cluster_one
+
 package k8sgateway_test
 
 import (
