+++ test/kubernetes/e2e/tests/classic_api_test.go
@@ -1,4 +1,6 @@
-package example
+//go:build cluster_one
+
+package tests_test
 
 import (
 	"context"
@@ -7,24 +9,25 @@ import (
 	"time"
 
 	"github.com/solo-io/gloo/test/kube2e/helper"
+	"github.com/solo-io/gloo/test/kubernetes/e2e/features/headless_svc"
 
 	"github.com/solo-io/skv2/codegen/util"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/solo-io/gloo/test/kubernetes/e2e"
-	"github.com/solo-io/gloo/test/kubernetes/e2e/features/example"
 	"github.com/solo-io/gloo/test/kubernetes/testutils/gloogateway"
 )
 
-// TestInstallationWithDebugLogLevel is the function which executes a series of tests against a given installation
-func TestInstallationWithDebugLogLevel(t *testing.T) {
+// TestClassicEdgeGateway is the function which executes a series of tests against a given installation where
+// the k8s Gateway controller is disabled
+func TestClassicEdgeGateway(t *testing.T) {
 	ctx := context.Background()
 	testCluster := e2e.MustTestCluster()
 	testInstallation := testCluster.RegisterTestInstallation(
 		t,
 		&gloogateway.Context{
-			InstallNamespace:   "debug-example",
-			ValuesManifestFile: filepath.Join(util.MustGetThisDir(), "manifests", "debug-example.yaml"),
+			InstallNamespace:   "classic-edge-test",
+			ValuesManifestFile: filepath.Join(util.MustGetThisDir(), "manifests", "classic-gateway-test-helm.yaml"),
 		},
 	)
 
@@ -43,11 +46,12 @@ func TestInstallationWithDebugLogLevel(t *testing.T) {
 		testCluster.UnregisterTestInstallation(testInstallation)
 	})
 
+	// Install Gloo Gateway with only classic APIs enabled
 	testInstallation.InstallGlooGateway(ctx, func(ctx context.Context) error {
 		return testHelper.InstallGloo(ctx, helper.GATEWAY, 5*time.Minute, helper.ExtraArgs("--values", testInstallation.Metadata.ValuesManifestFile))
 	})
 
-	t.Run("Example", func(t *testing.T) {
-		suite.Run(t, example.NewTestingSuite(ctx, testInstallation))
+	t.Run("HeadlessSvc", func(t *testing.T) {
+		suite.Run(t, headless_svc.NewClassicHeadlessSvcSuite(ctx, testInstallation))
 	})
 }
