+++ test/kubernetes/e2e/tests/debug_logging_test.go
@@ -1,4 +1,6 @@
-package classicedge_test
+//go:build cluster_example
+
+package tests
 
 import (
 	"context"
@@ -7,25 +9,24 @@ import (
 	"time"
 
 	"github.com/solo-io/gloo/test/kube2e/helper"
-	"github.com/solo-io/gloo/test/kubernetes/e2e/features/headless_svc"
 
 	"github.com/solo-io/skv2/codegen/util"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/solo-io/gloo/test/kubernetes/e2e"
+	"github.com/solo-io/gloo/test/kubernetes/e2e/features/example"
 	"github.com/solo-io/gloo/test/kubernetes/testutils/gloogateway"
 )
 
-// TestClassicEdgeGateway is the function which executes a series of tests against a given installation where
-// the k8s Gateway controller is disabled
-func TestClassicEdgeGateway(t *testing.T) {
+// TestInstallationWithDebugLogLevel is the function which executes a series of tests against a given installation
+func TestInstallationWithDebugLogLevel(t *testing.T) {
 	ctx := context.Background()
 	testCluster := e2e.MustTestCluster()
 	testInstallation := testCluster.RegisterTestInstallation(
 		t,
 		&gloogateway.Context{
-			InstallNamespace:   "classic-edge-test",
-			ValuesManifestFile: filepath.Join(util.MustGetThisDir(), "manifests", "classic-gateway-test-helm.yaml"),
+			InstallNamespace:   "debug-example",
+			ValuesManifestFile: filepath.Join(util.MustGetThisDir(), "manifests", "debug-example.yaml"),
 		},
 	)
 
@@ -44,12 +45,11 @@ func TestClassicEdgeGateway(t *testing.T) {
 		testCluster.UnregisterTestInstallation(testInstallation)
 	})
 
-	// Install Gloo Gateway with only classic APIs enabled
 	testInstallation.InstallGlooGateway(ctx, func(ctx context.Context) error {
 		return testHelper.InstallGloo(ctx, helper.GATEWAY, 5*time.Minute, helper.ExtraArgs("--values", testInstallation.Metadata.ValuesManifestFile))
 	})
 
-	t.Run("HeadlessSvc", func(t *testing.T) {
-		suite.Run(t, headless_svc.NewClassicHeadlessSvcSuite(ctx, testInstallation))
+	t.Run("Example", func(t *testing.T) {
+		suite.Run(t, example.NewTestingSuite(ctx, testInstallation))
 	})
 }
