+++ test/kubernetes/e2e/tests/manifests/debug-example.yaml
