+++ test/kubernetes/e2e/tests/manifests/error-example.yaml
